+++ 기능목록.md
@@ -16,6 +16,7 @@
         -   [x] card-number: 4글자까지
         -   [x] exp-month: 2글자까지, 1에서 12 사이
         -   [x] cvc: 3글자까지
+    -   유효성 검사를 Input 컴포넌트에서 알 필요가 없는 것 같아서 Input은 value 관리 및 스타일만 담당하고 유효성 검사는 외부(InputWrapper)에서 실시하는 걸로
 
 -   카드 브랜드 구분 로직
